backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis and Celery
redis==5.0.1
celery==5.3.4

# Authentication
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
requests==2.31.0

# Pydantic settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# Monitoring
prometheus-client==0.19.0

# Email
aiosmtplib==3.0.1

# Testing (optional, for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
ruff==0.1.6
